Move BrowserMode and BrowserModeStateHolder to the browser-api module
